32th test has failed

+++ internal/domain/models/status.go
@@ -2,8 +2,8 @@ package models
 
 //easyjson:json
 type Status struct {
-	Forum  int64 `json:"forum"`
-	Post   int64 `json:"post"`
-	Thread int64 `json:"thread"`
-	User   int64 `json:"user"`
+	Forum  int `json:"forum"`
+	Post   int `json:"post"`
+	Thread int `json:"thread"`
+	User   int `json:"user"`
 }
